delete nonusage method
