+++ src/main/java/softeer2nd/Main.java
@@ -2,6 +2,7 @@ package softeer2nd;
 
 import softeer2nd.chess.Board;
 import softeer2nd.chess.game.GameChess;
+import softeer2nd.chess.view.GameView;
 
 import java.io.*;
 import java.util.StringTokenizer;
@@ -13,11 +14,12 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         Board board = new Board();
         GameChess gameChess = board.getGameChess();
+        GameView gameView = board.getGameView();
         String str = st.nextToken();
 
 
         if (str.equals("시작")) {
-            board.print();
+            gameView.print();
             while(true){
                 st = new StringTokenizer(br.readLine());
                 String command = st.nextToken();
@@ -25,7 +27,7 @@ public class Main {
                     String from = st.nextToken();
                     String to = st.nextToken();
                     gameChess.move(from, to);
-                    board.print();
+                    gameView.print();
                 } else if (command.equals("종료")) {
                     break;
                 }
